requirements.txt
# requirements.txt
requests==2.31.0
twilio==8.3.0
python-dotenv==1.0.0